+++ eras/conway/impl/CHANGELOG.md
@@ -45,6 +45,7 @@
   * `ConwayCommitteeCert` -> `ConwayGovCert`
   * `ConwayTxCertCommittee` -> `ConwayTxCertGov`
 * Remove `DelegStakeTxCert` from the `COMPLETE` pragma for `TxCert`
+* Add `Committee` and adjust `NewCommittee` governance action
 
 ## 1.6.3.0